Cleanup for converted integration files
Finished converting src library

Rearranged files

Removed coffeescript, added CHANGELOG.md entry

Cleanup

Second round cleanup

Pull Request Fixes

IIFEs be gone
