+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.18.3] - 2020-03-27
+### Fixed
+- Decaffeinated the integration
+
 ## [2.18.2] - 2020-02-19
 ### Fixed
 - 'price' is now parsed regardless of presence of success vars
