+++ lib/form.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS104: Avoid inline assignments
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const flat = require('flat');
 const querystring = require('querystring');
@@ -26,16 +17,15 @@ const headers = require('./headers');
 const request = function(vars) {
 
   // user-defined form field values
-  let body, left;
   const formFields = vars.form_field != null ? vars.form_field : {};
-
-  const encodeValuesOnly = (vars.encode_form_field_names != null) && !(vars.encode_form_field_names != null ? vars.encode_form_field_names.valueOf() : undefined);
-
+  const encodeValuesOnly = (vars.encode_form_field_names != null) && !(_.result(vars, 'encode_form_fields.valueOf'));
+  
   // build body content
-  const normalized = normalize(formFields, (left = (vars.send_ascii != null ? vars.send_ascii.valueOf() : undefined)) != null ? left : false, encodeValuesOnly);
+  const normalized = normalize(formFields, _.result(vars, 'send_ascii.valueOf', false), encodeValuesOnly);
   const content = flat.flatten(compact(normalized), {safe: true});
-
+  
   // URL encode post body
+  let body;
   if (encodeValuesOnly) {
     body = querystring.stringify(content, null, null, {encodeURIComponent: component => { return component; } });
   } else {
@@ -73,11 +63,9 @@ request.variables = () => [
 //
 
 module.exports = {
-  name: 'Form POST',
   request,
   response,
   validate(vars) {
-    let left, left1;
-    return (left = (left1 = validate.url(vars)) != null ? left1 : validate.outcome(vars)) != null ? left : validate.headers(vars, 'form-urlencoded');
+    return validate.url(vars) || validate.outcome(vars) || validate.headers(vars, 'form-urlencoded');
   }
 };
