+++ lib/json.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS104: Avoid inline assignments
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const helpers = require('./helpers');
 const response = require('./response');
@@ -23,9 +14,7 @@ const compact = require('./compact');
 //
 
 const request = function(vars) {
-
-  let left, left1, left2;
-  let json = (left = normalize(vars.json_property, (left1 = (vars.send_ascii != null ? vars.send_ascii.valueOf() : undefined)) != null ? left1 : false)) != null ? left : {};
+  let json = normalize(vars.json_property, _.result(vars, 'send_ascii.valueOf', false)) || {}; 
 
   // test whether the mappings indicate that a root array is being requested
   const keys = Object.keys(json);
@@ -59,7 +48,7 @@ const request = function(vars) {
 
   return {
     url: vars.url,
-    method: (left2 = (vars.method != null ? vars.method.toUpperCase() : undefined)) != null ? left2 : 'POST',
+    method: _.result(vars, 'method.toUpperCase', 'POST'),
     headers: _.merge(defaultHeaders, headers(vars.header)),
     body
   };
@@ -86,12 +75,13 @@ request.variables = () => [
 //
 
 module.exports = {
-  name: 'JSON',
   request,
   response,
   validate(vars) {
-    let left, left1, left2;
-    return (left = (left1 = (left2 = validate.url(vars)) != null ? left2 : validate.method(vars, ['POST', 'PUT', 'DELETE'])) != null ? left1 : validate.outcome(vars)) != null ? left : validate.headers(vars, 'json');
+    return validate.url(vars) ||
+           validate.method(vars, ['POST', 'PUT', 'DELETE']) || 
+           validate.outcome(vars) ||
+           validate.headers(vars, 'json') ||
+           undefined;
   }
 };
-
