+++ lib/normalize.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
-let normalize;
 const _ = require('lodash');
 const flat = require('flat');
 const unidecode = require('unidecode');
@@ -14,12 +5,13 @@ const querystring = require('querystring');
 
 // use valueOf to ensure the normal version is sent for all richly typed values
 const valueOf = function(value, toAscii) {
-  if (value == null) { return value; }
-  if ((value != null ? value.valid : undefined) === false) {
+  if (!value) return value;
+
+  if (_.get(value, 'valid') == false) {
     // invalid richly typed values should be set to the raw value
     if (toAscii) { return unidecode(value.raw); } else { return value.raw; }
   } else {
-    if ((value != null ? value.valueOf() : undefined) != null) {
+    if (_.result(value, 'valueOf') != null) {
       if (toAscii) { return unidecode(value.valueOf()); } else { return value.valueOf(); }
     } else {
       return null;
@@ -28,10 +20,8 @@ const valueOf = function(value, toAscii) {
 };
 
 
-module.exports = (normalize = function(obj, toAscii, encodeValues) {
-  if (toAscii == null) { toAscii = false; }
-  if (encodeValues == null) { encodeValues = false; }
-  if (obj == null) { return obj; }
+const normalize = function(obj, toAscii = false, encodeValues = false) {
+  if (!obj) { return obj; }
 
   if (_.isArray(obj)) {
     return obj.map(val => normalize(val, toAscii));
@@ -48,4 +38,6 @@ module.exports = (normalize = function(obj, toAscii, encodeValues) {
   } else {
     return valueOf(obj, toAscii);
   }
-});
+};
+
+module.exports = normalize;
