+++ lib/query.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS104: Avoid inline assignments
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const flat = require('flat');
 const querystring = require('querystring');
@@ -25,11 +16,10 @@ const compact = require('./compact');
 const request = function(vars) {
 
   // user-defined form field values
-  let left;
   const parameters = vars.parameter != null ? vars.parameter : {};
 
   // build query content
-  const content = flat.flatten(normalize(parameters, (left = (vars.send_ascii != null ? vars.send_ascii.valueOf() : undefined)) != null ? left : false), {safe: true});
+  const content = flat.flatten(normalize(parameters, _.result(vars, 'send_ascii.valueOf', false)), {safe: true});
 
   // URL encoded post body
   const query = querystring.encode(compact(content));
@@ -57,12 +47,16 @@ request.variables = () => [
 // Exports ----------------------------------------------------------------
 //
 
+//
+// Validate --------------------------------------------------------------
+//
+
 module.exports = {
-  name: 'GET Query',
   request,
   response,
-  validate(vars) {
-    let left, left1;
-    return (left = (left1 = validate.url(vars)) != null ? left1 : validate.outcome(vars)) != null ? left : validate.headers(vars, 'No_Content_Type_allowed_for_GET');
+  validate: function(vars) {
+    return validate.url(vars) || 
+           validate.outcome(vars) ||
+           validate.headers(vars, 'No_Content_Type_allowed_for_GET')
   }
 };
