+++ lib/soap.js
@@ -1,14 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS103: Rewrite code to no longer use __guard__
- * DS104: Avoid inline assignments
- * DS205: Consider reworking code to avoid use of IIFEs
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const flat = require('flat');
 const soap = require('soap');
@@ -19,25 +8,14 @@ const normalize = require('./normalize');
 const validate = require('./validate');
 const compact = require('./compact');
 
-const helpers = require('./helpers');
-const {
-  ensureArray
-} = helpers;
-const {
-  inverseOutcome
-} = helpers;
-const {
-  toRegex
-} = helpers;
-
-
+const { ensureArray, inverseOutcome, toRegex } = require('./helpers');
 
 const handle = function(vars, callback) {
 
   vars = flat.unflatten(vars);
   const options = {
     valueKey: '#value',
-    forceSoap12Headers: (vars.version != null ? vars.version.trim() : undefined) === '1.2',
+    forceSoap12Headers: (vars.version ? vars.version.trim() : undefined) === '1.2',
     disableCache: true
   };
 
@@ -53,8 +31,7 @@ const handle = function(vars, callback) {
     ]);
   }
 
-  return soap.createClient(vars.url, options, function(err, client) {
-    let left;
+  soap.createClient(vars.url, options, function(err, client) {
     if (err) { return callback(err); }
 
     // the SOAP function to execute
@@ -75,7 +52,10 @@ const handle = function(vars, callback) {
     };
 
     // get SOAP function arguments
-    let args = normalize(vars.arg != null ? vars.arg : {}, (left = (vars.send_ascii != null ? vars.send_ascii.valueOf() : undefined)) != null ? left : false);
+    let args = normalize(
+      vars.arg || {}, 
+      _.result(vars, 'send_ascii.valueOf', false)
+    )
 
     // This routine converts each argument that is an object to an XML string, if the argument's type calls for it.
     args = _.mapValues(args, function(value, key) {
@@ -125,23 +105,20 @@ const handle = function(vars, callback) {
 
     // the callback to handle the SOAP function response
     const handleResponse = function(err, result, body) {
-      let left1;
-      const fault = __guard__(__guard__(__guard__(err != null ? err.root : undefined, x2 => x2.Envelope), x1 => x1.Body), x => x.Fault);
+      const fault = _.get(err, 'root.Envelope.Body.Fault', null)
       if (fault != null) {
-        return callback(null, {outcome: 'error', reason: fault.faultstring != null ? fault.faultstring : fault.faultcode});
+        return callback(null, { outcome: 'error', reason: fault.faultstring || fault.faultcode });
       } else if (err) {
         return callback(err);
       }
 
-      if (result == null) { result = {}; }
-
+      result = result || {};
       // Sometimes, SOAP functions return a string that contains encoded XML.
       // This routine tries to detect that scenario so that the XML can be parsed.
       result = _.mapValues(result, function(value) {
         if (_.isString(value)) {
-          let left1, left2;
-          const openElementCount = ((left1 = value.match(/</g)) != null ? left1 : []).length;
-          const closeElementCount = ((left2 = value.match(/>/g)) != null ? left2 : []).length;
+          const openElementCount = _.get(value.match(/</g), 'length', 0);
+          const closeElementCount = _.get(value.match(/>/g), 'length', 0);
           if ((openElementCount > 0) && (openElementCount === closeElementCount)) {
             // This seems like a well-formed XML string, so parse it as such
             const opts = {
@@ -150,49 +127,39 @@ const handle = function(vars, callback) {
               mergeAttrs: true,
               trim: true
             };
-            try { return mimecontent(value, 'text/xml').toObject(opts); } catch (err) {}
+            try {
+              return mimecontent(value, 'text/xml').toObject(opts);
+            } catch (err) {}
           }
         }
         return value;
       });
 
       const searchTerm = toRegex(vars.outcome_search_term);
-      const searchOutcome = (left1 = (vars.outcome_on_match != null ? vars.outcome_on_match.trim().toLowerCase() : undefined)) != null ? left1 : 'success';
-      const searchPath = vars.outcome_search_path != null ? vars.outcome_search_path.trim() : undefined;
-      const reasonSelector = vars.reason_path != null ? vars.reason_path.trim() : undefined;
-      const priceSelector = vars.price_path != null ? vars.price_path.trim() : undefined;
+      const searchOutcome  = _.result(vars, 'outcome_on_match.trim', 'success').toLowerCase()
+      const searchPath     = _.result(vars, 'outcome_search_path.trim');
+      const reasonSelector = _.result(vars, 'reason_path.trim');
+      const priceSelector  = _.result(vars, 'price_path.trim');
 
       // narrow the search scope
+      
       const searchIn =
         searchPath ?
           // limit outcome_search_term to this path in the document
-          __guard__(_.get(result, searchPath), x3 => x3.toString())
-
+          _.get(result, searchPath).toString()
         :
           // no search path was provided, so search the entire response body
           body;
 
       // look in all possible search scopes for the term
-      const found = searchIn != null ? searchIn.toLowerCase().match(searchTerm) : undefined;
+      const found = (searchIn) ? searchIn.toLowerCase().match(searchTerm) : undefined;
 
       // determine the outcome based on whether the search term was found
-      const outcome =
-        found ?
-          searchOutcome
-        :
-          inverseOutcome(searchOutcome);
-
-      const price = 
-        outcome === 'success' ?
-          _.get(result, priceSelector) : undefined;
-
+      const outcome = (found) ? searchOutcome : inverseOutcome(searchOutcome);
+      const price = (outcome === 'success') ? _.get(result, priceSelector) : undefined;
 
       // determine the reason based on the reason selector
-      const reasons =
-        reasonSelector ?
-          _.get(result, reasonSelector)
-        :
-          [];
+      const reasons = (reasonSelector) ? _.get(result, reasonSelector) : [];
 
       // trim and comma delimit reasons
       let reason = _(ensureArray(reasons))
@@ -202,7 +169,7 @@ const handle = function(vars, callback) {
         .join(', ');
 
       // set the default reason, if necessary
-      if (!reason) { reason = vars.default_reason != null ? vars.default_reason.trim() : undefined; }
+      if (!reason) { reason = _.result(vars, 'default_reason.trim') }
 
       // build the event
       const event = result;
@@ -211,14 +178,14 @@ const handle = function(vars, callback) {
       event.price = price || 0;
 
       // return the event
-      return callback(null, event);
+      callback(null, event);
     };
 
     // calculate the timeout
     const timeoutMs = (vars.timeout_seconds != null ? vars.timeout_seconds : 10) * 1000;
 
     // call the function
-    return fxn(args, handleResponse, {timeout: timeoutMs});
+    fxn(args, handleResponse, {timeout: timeoutMs});
   });
 };
 
@@ -229,14 +196,14 @@ const handle = function(vars, callback) {
 //
 
 const validateFunction = function(vars) {
-  if (vars.function == null) { return 'Function is required'; }
+  if (!vars.function) { return 'Function is required'; }
   if (!vars.function.match(/^[a-zA-Z0-9_.]+$/)) { return 'Function must have valid name'; }
 };
 
 
 const validateVersion = function(vars) {
   const version = vars.version != null ? vars.version.trim() : undefined;
-  if (version == null) { return; }
+  if (!version) { return; }
   if ((version !== '1.1') && (version !== '1.2')) { return 'Must be valid SOAP version: 1.1 or 1.2'; }
 };
 
@@ -281,18 +248,13 @@ const responseVariables = () => [
 
 
 module.exports = {
-  name: 'SOAP',
   handle,
   requestVariables,
   responseVariables,
   validate(vars) {
-    let left, left1, left2;
-    return (left = (left1 = (left2 = validate.url(vars)) != null ? left2 : validate.outcome(vars)) != null ? left1 : validateFunction(vars)) != null ? left : validateVersion(vars);
+    return validate.url(vars) ||
+           validate.outcome(vars) ||
+           validateFunction(vars) ||
+           validateVersion(vars)
   }
 };
-
-
-
-function __guard__(value, transform) {
-  return (typeof value !== 'undefined' && value !== null) ? transform(value) : undefined;
-}
