+++ lib/xml-doc.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS104: Avoid inline assignments
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const flat = require('flat');
 const normalize = require('./normalize');
@@ -17,11 +8,8 @@ const compact = require('./compact');
 // Convert an object into one that can be used by xmlbuilder to create a XML document.
 // Keys containing @ are treated as attributes.
 //
-module.exports = function(obj, toAscii) {
-
-  let left;
-  if (toAscii == null) { toAscii = false; }
-  let xmlPaths = flat.flatten((left = normalize(obj, toAscii)) != null ? left : {});
+module.exports = function(obj, toAscii = false) {
+  let xmlPaths = flat.flatten(normalize(obj, toAscii) || {});
 
   xmlPaths =
     _(xmlPaths)
