+++ lib/xml.js
@@ -1,12 +1,3 @@
-// TODO: This file was created by bulk-decaffeinate.
-// Sanity-check the conversion and remove this comment.
-/*
- * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
- * DS104: Avoid inline assignments
- * DS207: Consider shorter variations of null checks
- * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
- */
 const _ = require('lodash');
 const builder = require('xmlbuilder');
 const helpers = require('./helpers');
@@ -23,16 +14,9 @@ const headers = require('./headers');
 //
 
 const request = function(vars) {
-
-  let left, left1;
-  const obj = xmlDoc(vars.xml_path, (left = (vars.send_ascii != null ? vars.send_ascii.valueOf() : undefined)) != null ? left : false);
-  let body =
-    Object.keys(obj).length ?
-      builder.create(obj).end({pretty: true})
-    :
-      '<?xml version="1.0"?>';
-
-  let contentType = (vars.header != null ? vars.header['Content-Type'] : undefined) != null ? (vars.header != null ? vars.header['Content-Type'] : undefined) : 'text/xml';
+  const obj = xmlDoc(vars.xml_path, _.result(vars, 'send_ascii.valueOf', false));
+  let body = (Object.keys(obj).length) ? builder.create(obj).end({pretty: true}) : '<?xml version="1.0"?>';
+  let contentType = (vars.header && vars.header['Content-Type']) ? vars.header['Content-Type'] : 'text/xml';
 
   if (vars.xml_parameter) {
     body = helpers.parameterize(vars.xml_parameter, body, vars.extra_parameter);
@@ -41,7 +25,7 @@ const request = function(vars) {
 
   return {
     url: vars.url,
-    method: (left1 = (vars.method != null ? vars.method.toUpperCase() : undefined)) != null ? left1 : 'POST',
+    method: _.result(vars, 'method.toUpperCase') || 'POST',
     headers: _.merge(headers(vars.header), {
       'Content-Type': contentType,
       'Content-Length': Buffer.byteLength(body),
@@ -73,12 +57,12 @@ request.variables = () => [
 //
 
 module.exports = {
-  name: 'XML',
   request,
   response,
   validate(vars) {
-    let left, left1, left2;
-    return (left = (left1 = (left2 = validate.url(vars)) != null ? left2 : validate.method(vars, ['POST', 'PUT'])) != null ? left1 : validate.outcome(vars)) != null ? left : validate.headers(vars, 'xml');
+    return validate.url(vars) ||
+           validate.method(vars, ['POST', 'PUT']) ||
+           validate.outcome(vars) ||
+           validate.headers(vars, 'xml')
   }
 };
-
