+++ package.json
@@ -2,7 +2,7 @@
   "name": "leadconduit-custom",
   "description": "LeadConduit custom integration",
   "author": "Alex Wolfe <alex@activeprospect.com>",
-  "version": "2.18.4",
+  "version": "2.18.3",
   "main": "index.js",
   "license": "CC-BY-NC-ND-4.0",
   "repository": {
@@ -10,9 +10,7 @@
     "url": "https://github.com/activeprospect/leadconduit-integration-custom.git"
   },
   "scripts": {
-    "test": "cake test",
-    "jstest": "mocha",
-    "prepublish": "cake build"
+    "test": "mocha"
   },
   "dependencies": {
     "dot-wild": "^3.0.1",
@@ -29,8 +27,6 @@
   },
   "devDependencies": {
     "chai": ">= 1.9.0",
-    "coffee-script": ">= 1.7.0",
-    "leadconduit-cakefile": "^0.3.1",
     "leadconduit-types": "^4.5.0",
     "mocha": "^7.1.1",
     "nock": "^12.0.0"
