--- spec/soap-failure.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<soap:Envelope xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/">
-    <soap:Body>
-        <AddLeadResponse xmlns="http://donkey/ws.asmx/">
-            <AddLeadResult>
-                <Result>failure</Result>
-                <Message>duplicate</Message>
-                <LeadId>12346</LeadId>
-            </AddLeadResult>
-        </AddLeadResponse>
-    </soap:Body>
-</soap:Envelope>
