--- spec/soap-success.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<soap:Envelope xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/">
-    <soap:Body>
-        <AddLeadResponse xmlns="http://donkey/ws.asmx/">
-            <AddLeadResult>
-                <Result>true</Result>
-                <Message>some message</Message>
-                <LeadId>12345</LeadId>
-                <Cost>1.5</Cost>
-                <Empty/>
-                <Multi>
-                    <Foo>1</Foo>
-                    <Foo>2</Foo>
-                </Multi>
-            </AddLeadResult>
-        </AddLeadResponse>
-    </soap:Body>
-</soap:Envelope>
