+++ test/compact-spec.js
@@ -1,6 +1,4 @@
-const {
-  assert
-} = require('chai');
+const { assert } = require('chai');
 const compact = require('../lib/compact');
 
 describe('Compact', function() {
@@ -17,6 +15,3 @@ describe('Compact', function() {
 
   it('should compact array nested in object property array', () => assert.deepEqual(compact({foo: [1, 2, { bar: [ 'a', null, 'c' ]}, 4]}), {foo: [1, 2, { bar: ['a', 'c'] }, 4]}));
 });
-
-
-
